refactor: separate EditableNode from FlowNode
