+++ web/src/modules/topic/components/Diagram/Diagram.tsx
@@ -7,7 +7,7 @@ import {
   BackgroundVariant,
   type EdgeProps as DefaultEdgeProps,
   type NodeProps as DefaultNodeProps,
-  type Node as FlowNode,
+  type Node as FlowNodeState,
 } from "reactflow";
 
 import { ContextMenu } from "../../../../common/components/Menu/Menu";
@@ -16,14 +16,14 @@ import { closeClaimDiagram, connectNodes, deselectNodes } from "../../store/acti
 import { useFilteredDiagram } from "../../store/store";
 import { type Edge, type Node } from "../../utils/diagram";
 import { type NodeType } from "../../utils/nodes";
-import { EditableNode } from "../EditableNode/EditableNode";
+import { FlowNode } from "../FlowNode/FlowNode";
 import { ScoreEdge } from "../ScoreEdge/ScoreEdge";
 import { PositionedIconButton, StyledReactFlow } from "./Diagram.styles";
 
 const buildNodeComponent = (type: NodeType) => {
   // eslint-disable-next-line react/display-name -- react flow dynamically creates these components without name anyway
   return (props: NodeProps) => {
-    return <EditableNode {...props} type={type} />;
+    return <FlowNode {...props} type={type} />;
   };
 };
 
@@ -84,7 +84,9 @@ export const Diagram = ({ diagramId }: DiagramProps) => {
         minZoom={0.25}
         onPaneClick={deselectNodes}
         onConnect={({ source, target }) => connectNodes(source, target)}
-        onNodeContextMenu={(event: MouseEvent, node: FlowNode) => openContextMenu(event, { node })}
+        onNodeContextMenu={(event: MouseEvent, node: FlowNodeState) =>
+          openContextMenu(event, { node })
+        }
         nodesDraggable={false}
         nodesConnectable={diagram.type !== "claim"} // claim diagram is a tree, so cannot connect existing nodes
       >
