+++ web/src/modules/topic/components/EditableNode/EditableNode.tsx
@@ -1,65 +1,41 @@
-import { Global } from "@emotion/react";
 import { useTheme } from "@mui/material";
 import _ from "lodash";
-import { Handle, Position } from "reactflow";
 
 import { setNodeLabel } from "../../store/actions";
-import { useDiagramType } from "../../store/store";
-import { orientations } from "../../utils/diagram";
-import { NodeType, nodeDecorations } from "../../utils/nodes";
+import { Node } from "../../utils/diagram";
+import { nodeDecorations } from "../../utils/nodes";
 import { CriteriaIndicator } from "../CriteriaIndicator/CriteriaIndicator";
 import { CriteriaTableIndicator } from "../CriteriaTableIndicator/CriteriaTableIndicator";
-import { NodeProps } from "../Diagram/Diagram";
 import { ScoreDial } from "../ScoreDial/ScoreDial";
 import {
-  AddNodeButtonGroupChild,
-  AddNodeButtonGroupParent,
-  HoverBridgeDiv,
   IndicatorDiv,
   MiddleDiv,
+  NodeDiv,
   NodeTypeDiv,
   NodeTypeSpan,
   StyledTextareaAutosize,
   XEdgeDiv,
   YEdgeDiv,
-  nodeStyles,
 } from "./EditableNode.styles";
 
-export const EditableNode = ({ id, data, type }: NodeProps) => {
-  const diagramType = useDiagramType(data.diagramId);
+export const EditableNode = ({ node }: { node: Node }) => {
   const theme = useTheme();
 
-  if (!diagramType) return <></>;
-
-  const orientation = orientations[diagramType];
-
-  const nodeType = type as NodeType; // we always pass a NodeType from the diagram, but I'm not sure how to override react-flow's type to tell it that
-  const nodeDecoration = nodeDecorations[nodeType];
-  const color = theme.palette[nodeType].main;
+  const nodeDecoration = nodeDecorations[node.type];
+  const color = theme.palette[node.type].main;
   const NodeIcon = nodeDecoration.NodeIcon;
 
   return (
-    <>
-      <HoverBridgeDiv />
-
-      <Handle type="target" position={orientation == "TB" ? Position.Top : Position.Left} />
-      {/* should this use react-flow's NodeToolbar? seems like it'd automatically handle positioning */}
-      <AddNodeButtonGroupParent
-        fromNodeId={id}
-        fromNodeType={nodeType}
-        as="parent"
-        orientation={orientation}
-      />
-
+    <NodeDiv color={color} className={node.selected ? "selected" : ""}>
       <YEdgeDiv>
         <NodeTypeDiv>
           <NodeIcon sx={{ width: "16px", height: "16px" }} />
-          <NodeTypeSpan>{_.startCase(nodeType)}</NodeTypeSpan>
+          <NodeTypeSpan>{_.startCase(node.type)}</NodeTypeSpan>
         </NodeTypeDiv>
         <IndicatorDiv>
-          <CriteriaTableIndicator nodeId={id} diagramId={data.diagramId} />
-          <CriteriaIndicator nodeId={id} diagramId={data.diagramId} />
-          <ScoreDial scorableId={id} scorableType="node" score={data.score} />
+          <CriteriaTableIndicator nodeId={node.id} diagramId={node.data.diagramId} />
+          <CriteriaIndicator nodeId={node.id} diagramId={node.data.diagramId} />
+          <ScoreDial scorableId={node.id} scorableType="node" score={node.data.score} />
         </IndicatorDiv>
       </YEdgeDiv>
       <MiddleDiv>
@@ -71,24 +47,14 @@ export const EditableNode = ({ id, data, type }: NodeProps) => {
           // Was previously using defaultValue to avoid this, but that caused text to not update
           // when rendering for the second time (1. post-hydration value updating, see store, or
           // 2. when importing a new diagram but the node id's are the same).
-          value={data.label}
+          value={node.data.label}
           maxRows={3}
-          onChange={(event) => setNodeLabel(id, event.target.value)}
+          onChange={(event) => setNodeLabel(node.id, event.target.value)}
           className="nopan" // allow regular text input drag functionality without using reactflow's pan behavior
         />
         <XEdgeDiv />
       </MiddleDiv>
       <YEdgeDiv />
-
-      <AddNodeButtonGroupChild
-        fromNodeId={id}
-        fromNodeType={nodeType}
-        as="child"
-        orientation={orientation}
-      />
-      <Handle type="source" position={orientation == "TB" ? Position.Bottom : Position.Right} />
-
-      <Global styles={nodeStyles(data.width, color, nodeType)} />
-    </>
+    </NodeDiv>
   );
 };
